+++ src/react-app/src/App.test.tsx
@@ -1,8 +1,0 @@
-import { render } from '@testing-library/react';
-import App from './App';
-
-test('renders component', () => {
-  const { getByText } = render(<App />);
-  const textElement = getByText('closeWindowMsg');
-  expect(textElement).toBeInTheDocument();
-});