+++ src/react-app/src/App.tsx
@@ -1,16 +1,15 @@
+import { GuxIcon } from 'genesys-spark-components-react';
 import './App.css';
 import { useTranslation } from 'react-i18next';
+import { handleRedirectFromLogin } from '../../lib/parse-redirect';
 
-declare var GenesysCloudClientAuth: {
-  handleRedirectFromLogin: () => void;
-};
-
-export default function App() {
+const App = () => {
   const { t } = useTranslation();
   let response = <h3>{t('successMsg')}</h3>;
 
   try {
-    GenesysCloudClientAuth.handleRedirectFromLogin();
+    handleRedirectFromLogin();
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   } catch (e: any) {
     response = (
       <div>
@@ -23,14 +22,16 @@ export default function App() {
   return (
     <div className='App'>
       <div className='content'>
-        <gux-icon
+        <GuxIcon
           className='logo'
-          icon-name='genesys-logo-full'
+          icon-name='legacy/genesys-logo-full'
           decorative={true}
-        ></gux-icon>
-        {response}
+        ></GuxIcon>
+        { response }
         <h1>{t('closeWindowMsg')}</h1>
       </div>
     </div>
   );
 }
+
+export default App;